Fixed a bug in the cache_snoop module. Resolves issue #123.
Added 10 new profiler sites. (@WebBreacher)

Fixed a parsing bug in the facebook module. (@qkaiser)

Fixed a bug in the search_contacts jigsaw module. Resolves issue #125.

Added a threading API to the core module class.

Added a lock to the print override function to ensure thread safe output to the console and spool file.

Made a slight syntax change to the instagram pushpin module.

Fixed a bug that was caused by a previous name change to a framework method.

Added threading to the namechk module. (@PaulSec)

Removed some unncessary code from the namechk module.

Added the print_exception method for handling the display of exception information.

Added the threads global option to universally set the number of threads the framework will use for thread capable modules.

Updated the leaks_dump module to account for changes in the PwnedList API.

Updated the requests library to ignore SSL certificate validation.

Added threading to the profiler module. (@WebBreacher)